Created forms folder. Login button redirects to homepage.
