+++ forms/login.js
@@ -1,0 +1,89 @@
+import { StatusBar } from 'expo-status-bar';
+import React, { useState } from 'react';
+import { Text, View, Image, TextInput, Button, TouchableOpacity, StyleSheet} from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+
+export default function LoginScreen({ navigation }) {
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+  
+    return (
+      <View style={styles.container}>
+        <Image style={styles.image} source={require('../resources/MovieMatchLogo.png')} />
+   
+        <StatusBar style="auto" />
+        <View style={styles.inputView}>
+          <TextInput
+            style={styles.TextInput}
+            placeholder="Email"
+            placeholderTextColor="#003f5c"
+            onChangeText={(email) => setEmail(email)}
+          />
+        </View>
+   
+        <View style={styles.inputView}>
+          <TextInput
+            style={styles.TextInput}
+            placeholder="Password"
+            placeholderTextColor="#003f5c"
+            secureTextEntry={true}
+            onChangeText={(password) => setPassword(password)}
+          />
+        </View>
+   
+        <TouchableOpacity>
+          <Text style={styles.forgot_button}>Forgot Password?</Text>
+        </TouchableOpacity>
+   
+        <TouchableOpacity style={styles.loginBtn} onPress={() => navigation.navigate('Home')}>
+          <Text style={styles.loginText}>LOGIN</Text>
+        </TouchableOpacity>
+      </View>
+    );
+}
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: "#fff",
+      alignItems: "center",
+      justifyContent: "center",
+    },
+   
+    image: {
+      marginBottom: 40,
+    },
+   
+    inputView: {
+      backgroundColor: "#FFC0CB",
+      borderRadius: 30,
+      width: "70%",
+      height: 45,
+      marginBottom: 20,
+   
+      alignItems: "center",
+    },
+   
+    TextInput: {
+      height: 50,
+      flex: 1,
+      padding: 10,
+      marginLeft: 20,
+    },
+   
+    forgot_button: {
+      height: 30,
+      marginBottom: 30,
+    },
+   
+    loginBtn: {
+      width: "80%",
+      borderRadius: 25,
+      height: 50,
+      alignItems: "center",
+      justifyContent: "center",
+      marginTop: 40,
+      backgroundColor: "#FF1493",
+    },
+  });
